==> src/version_2/4_sample_analysis.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
import os
from scipy.stats import mannwhitneyu, wilcoxon
from statsmodels.stats.multitest import multipletests
import math
from statannotations.Annotator import Annotator

pd.options.display.float_format = '{:.3f}'.format
pd.set_option('display.max_columns', None)

plt.rcParams['figure.dpi']=170

DPI=250

from list_vars import LIST_PROFILERS, DIR_FIGURES, RESULTS_DIR, POOLS, CONTROLS

# Biological sample analysis (part 2)

In this notebook we are going to do the analysis on the *biological* samples (POOL samples + controls).

One thing we want to check is how S and mode affect the robustness of our results. So, to do that we are going to compare the results using a comparative plot. The comparativ plot shows how many new variable are shown compared to the previous case. So, we can do a comparative increase on mode by fixing S, or on S by fixing the mode.

This analysis can be performed with many variables, and we are going to choose the following:
1) The number of detected species, across all samples and per individual sample (after NORM+ and cutting of species with more than 65% NaNs).
2) The number of detected differentially abundant species across the 4 comparisons.

With this in mind, we can later select one S and one mode and do the following analyses.

1) The importance of including the biological control samples to ensure that false positives are not considered.
2) The importance of normalizing the reads considering the biogical samples.
3) Plot the significantly differentially abundant species. 


## Observing the effect of $S$ on species retention for biological controls

rename_mapping = {
        'POOL1': 'RR1', 'POOL2': 'RR2', 'POOL3': 'RR3', 'POOL4': 'RR4',
        'POOL5': 'SP1', 'POOL6': 'SP2', 'POOL7': 'SP3', 'POOL8': 'SP4',
        'POOL9': 'HC1', 'POOL10': 'HC2', 'POOL11': 'HC3', 'POOL12': 'HC4',
    }

dict_retained_discarded_species = {'mode': [], 'S': [], 'sample': [], 'retained': [], 'discarded': []}

for mode in [3, 5, 7]:
    for S in [0, 1, 2, 3, 4, 5, 6, 7, 10, 15]:
        for sample in [f'POOL{i}' for i in range(1, 13)] + ['ACIDOLA', 'BLACTIS']:
            flags_file =  f'{RESULTS_DIR}/summary/{sample}_pass2_mode{mode}_taxgenus_S{S}_NORM+.flags.tsv'
            df_flags = pd.read_csv(flags_file, sep='\t').set_index('taxonomy_id')[['name', 'lineage', 'mean_norm']]
            
            dict_retained_discarded_species['mode'].append(mode)
            dict_retained_discarded_species['S'].append(S)
            dict_retained_discarded_species['sample'].append(rename_mapping[sample] if 'POOL' in sample else sample)
            dict_retained_discarded_species['retained'].append((df_flags['mean_norm'] == False).sum())
            dict_retained_discarded_species['discarded'].append((df_flags['mean_norm'] == True).sum())          

df_retained_discarded_species = pd.DataFrame(dict_retained_discarded_species)
df_retained_discarded_species['total'] = df_retained_discarded_species['retained'] + df_retained_discarded_species['discarded']

df_retained_discarded_species['percentage'] = 100 * df_retained_discarded_species['retained'] / df_retained_discarded_species['total']


df_total_counts = df_retained_discarded_species.groupby(['mode', 'sample'])['total'].max().unstack(level=0)
df_total_counts['increment_3_5'] = df_total_counts[5] / df_total_counts[3]
df_total_counts['increment_3_7'] = df_total_counts[7] / df_total_counts[3]


df_total_counts['% 3 in 5'] = 0.0        
df_total_counts['% 3 in 7'] = 0.0

S= 15
for sample in [f'POOL{i}' for i in range(1, 13)] + ['BLACTIS', 'ACIDOLA', ]:
    flags_file_3 =  f'{RESULTS_DIR}/summary/{sample}_pass2_mode3_taxgenus_S{S}_NORM+.flags.tsv'
    df_flags_3 = pd.read_csv(flags_file_3, sep='\t').set_index('taxonomy_id')[['name', 'lineage', 'mean_norm']]     

    flags_file_5 =  f'{RESULTS_DIR}/summary/{sample}_pass2_mode5_taxgenus_S{S}_NORM+.flags.tsv'
    df_flags_5 = pd.read_csv(flags_file_5, sep='\t').set_index('taxonomy_id')[['name', 'lineage', 'mean_norm']] 

    flags_file_7 =  f'{RESULTS_DIR}/summary/{sample}_pass2_mode7_taxgenus_S{S}_NORM+.flags.tsv'
    df_flags_7 = pd.read_csv(flags_file_7, sep='\t').set_index('taxonomy_id')[['name', 'lineage', 'mean_norm']]  



    if sample in rename_mapping:
        sample = rename_mapping[sample]
    
    df_total_counts.loc[sample, '% 3 in 5'] = 100 * len(set(df_flags_3.index) & set(df_flags_5.index)) / len(set(df_flags_3.index))        
    df_total_counts.loc[sample, '% 3 in 7'] = 100 * len(set(df_flags_3.index) & set(df_flags_7.index)) / len(set(df_flags_3.index))

df_total_counts

for column, name in zip(['total', 'retained', 'percentage'], ['Total genera', 'Retained genera',  'Retained percentage']):
    # Set up the grid for 14 samples in a 4x4 layout
    samples = df_retained_discarded_species['sample'].unique()  # Unique samples
    n_cols, n_rows = 7, 2  # Define the grid layout

    # Initialize the grid layout
    fig, axes = plt.subplots(n_rows, n_cols, figsize=(2*n_cols, 2*n_rows), sharex=True)

    # Placeholder to collect handles and labels for the legend
    handles, labels = None, None


    # Loop through each sample and create its subplot
    for i, sample in enumerate(samples):
        row, col = divmod(i, n_cols)  # Get the row and column index for the subplot
        ax = axes[row, col]

        # Filter the DataFrame for the current sample
        sample_data = df_retained_discarded_species[df_retained_discarded_species['sample'] == sample]

        # Plot retained species as a function of S with mode as hue
        lineplot = sns.lineplot(
            data=sample_data,
            x='S',
            y=column,
            hue='mode',
            marker='o',
            ax=ax,
            legend=(i == 6)  # Enable legend only for the first subplot
        )

        ax.set_title(sample)  # Title of the subplot is the sample name
        ax.set_xlabel("S")
        if col == 0:
            ax.set_ylabel(name)
        else:
            ax.set_ylabel("")

        ax.set_xticks([0,  2,  4,  6,  10, 15])

        if i==6:
            ax.legend(frameon=False, bbox_to_anchor=(1.05, 1), title='Mode')

        # Set a specific Y-axis limit for the last two samples
        if column == 'retained':
            if i in [len(samples) - 2, len(samples) - 1]:  # Last two samples
                ax.set_ylim(0, 4)

    # Remove extra subplots (for grid cells not needed)
    for j in range(i + 1, n_rows * n_cols):
        fig.delaxes(axes[j // n_cols, j % n_cols])


    # Adjust the main plot area to leave space for the legend
    plt.tight_layout()

    for format in ['png', 'tiff']: 
        plt.savefig(f'{RESULTS_DIR}/figures/paper/s_effect_{column}.{format}', dpi=DPI)

    plt.show()

## Plotting sample detection and statistacally differential species across modes and S

# Initialize an empty list to collect data rows
data_rows = []

# Iterate through modes and S values
for mode in [3, 5, 7]:
    for S in [0, 1, 2, 3, 4, 5, 6, 7, 10, 15]:
        # Read the data
        df_normpipe_retained_species = pd.read_csv(f'{RESULTS_DIR}/merged_counts/mode{mode}_S{S}_NORM+_retained.tsv', sep='\t')
        df_normpipe_discarded_normplus = pd.read_csv(f'{RESULTS_DIR}/merged_counts/mode{mode}_S{S}_NORM+_discarded_norm+.tsv', sep='\t')
        df_normpipe_discarded_common = pd.read_csv(f'{RESULTS_DIR}/merged_counts/mode{mode}_S{S}_NORM+_discarded_common.tsv', sep='\t')

        # Get the list of samples (assuming they are column names)
        samples = df_normpipe_retained_species.columns[3:]

        data_rows.append({
                'mode': mode,
                'S': S,
                'sample': 'ALL',
                'count_retained': len(df_normpipe_retained_species),
                'count_discarded_norm': len(df_normpipe_discarded_normplus),
                'count_discarded_all': len(df_normpipe_discarded_common)
            })
        
        for sample in samples:
            # Count non-NaN species for each dataframe and sample
            count_retained = df_normpipe_retained_species[sample].notna().sum()
            count_discarded_norm = df_normpipe_discarded_normplus[sample].notna().sum()
            count_discarded_all = df_normpipe_discarded_common[sample].notna().sum()

            # Collect the data as a dictionary
            data_rows.append({
                'mode': mode,
                'S': S,
                'sample': sample,
                'count_retained': count_retained,
                'count_discarded_norm': count_discarded_norm,
                'count_discarded_all': count_discarded_all
            })

# Convert the collected rows into a dataframe
df_stats_species_count = pd.DataFrame(data_rows)
df_stats_species_count['count_total'] = df_stats_species_count['count_retained'] + df_stats_species_count['count_discarded_norm'] + df_stats_species_count['count_discarded_all']
df_stats_species_count

df_stats_species_count[df_stats_species_count['sample'] == 'ALL']

# Initialize an empty list to collect data rows
data_rows = []

# Iterate through modes and S values
for mode in [3, 5, 7]:
    for S in [0, 1, 2, 3, 4, 5, 6, 7, 10, 15]:
        # Read the data
        df_pval_HCvsRR = pd.read_csv(f'{RESULTS_DIR}/differential_abundance/mode{mode}_S{S}_HCvsRR.tsv', sep='\t')
        df_pval_HCvsSP = pd.read_csv(f'{RESULTS_DIR}/differential_abundance/mode{mode}_S{S}_HCvsSP.tsv', sep='\t')
        df_pval_RRvsSP = pd.read_csv(f'{RESULTS_DIR}/differential_abundance/mode{mode}_S{S}_RRvsSP.tsv', sep='\t')
        df_pval_sex = pd.read_csv(f'{RESULTS_DIR}/differential_abundance/mode{mode}_S{S}_sex.tsv', sep='\t')

        # Get the list of samples (assuming they are column names)
        samples = df_normpipe_retained_species.columns[3:]

        data_rows.append({
                'mode': mode,
                'S': S,
                'count_HCvsRR': len(df_pval_HCvsRR[df_pval_HCvsRR['pval_MW'] < 0.05]),
                'count_HCvsSP': len(df_pval_HCvsSP[df_pval_HCvsSP['pval_MW'] < 0.05]),
                'count_RRvsSP': len(df_pval_RRvsSP[df_pval_RRvsSP['pval_MW'] < 0.05]),
                'count_sex': len(df_pval_sex[df_pval_sex['pval_MW'] < 0.05]),
                'species_HCvsRR': df_pval_HCvsRR[df_pval_HCvsRR['pval_MW'] < 0.05]['name'].values.tolist(),
                'species_HCvsSP': df_pval_HCvsSP[df_pval_HCvsSP['pval_MW'] < 0.05]['name'].values.tolist(),
                'species_RRvsSP': df_pval_RRvsSP[df_pval_RRvsSP['pval_MW'] < 0.05]['name'].values.tolist(),
                'species_sex': df_pval_sex[df_pval_sex['pval_MW'] < 0.05]['name'].values.tolist()
            })
        
      

# Convert the collected rows into a dataframe
df_stats_species_diffabundance = pd.DataFrame(data_rows)
df_stats_species_diffabundance['count_total'] = df_stats_species_diffabundance['count_HCvsRR'] + df_stats_species_diffabundance['count_HCvsSP'] + \
                                                df_stats_species_diffabundance['count_RRvsSP'] + df_stats_species_diffabundance['count_sex']
df_stats_species_diffabundance


## Why normalization with biological controls is relevant

mode = 3
S = 7

df_normpipe_retained_species = pd.read_csv(f'{RESULTS_DIR}/merged_counts/mode{mode}_S{S}_NORM+_retained.tsv', sep='\t')
df_normpipe_discarded_normplus = pd.read_csv(f'{RESULTS_DIR}/merged_counts/mode{mode}_S{S}_NORM+_discarded_norm+.tsv', sep='\t')
df_normpipe_discarded_common = pd.read_csv(f'{RESULTS_DIR}/merged_counts/mode{mode}_S{S}_NORM+_discarded_common.tsv', sep='\t')

print(len(df_normpipe_retained_species), len(df_normpipe_discarded_normplus), len(df_normpipe_discarded_common))

species = 'Janibacter'

display(df_normpipe_retained_species[df_normpipe_retained_species['name'] == species])
display(df_normpipe_discarded_normplus[df_normpipe_discarded_normplus['name'] == species])
display(df_normpipe_discarded_common[df_normpipe_discarded_common['name'] == species])

## Plotting discarded species

def annotation_pval(df, ax, groups, x_col, y_col):
    if len(groups) == 2:
        pairs = [(groups[0], groups[1])]

    annotator = Annotator(ax, pairs, data=df, x=x_col, y=y_col)
    annotator.configure(test='Mann-Whitney', text_format='simple', loc='outside', line_width=0.65, fontsize=8)
    annotator.apply_and_annotate()

def add_medians(df, ax, x_col, y_col, ysync=0.1, xsync=0.1):
    groups = df[x_col].unique()
    vals = [df[df[x_col] == group][y_col].values for group in groups]

    medians = [np.nanmedian(val) for val in vals]
    medianslog =  [np.log10(np.nanmedian(val) + 1) for val in vals]
    

    for i, median in enumerate(medians):
            ax.text(i + xsync, medianslog[i] + ysync, f"{medians[i]:.0f}", ha='left', va='bottom', fontsize=7)

    log2fc = np.log2(medians[1] / medians[0])

    if log2fc > 0:
        ax.text(-0.4, 5, f"log$_2$FC: {log2fc:.2f}", ha='left', va='bottom', fontsize=7)
    else:
        ax.text(1.5, 5, f"log$_2$FC: {log2fc:.2f}", ha='right', va='bottom', fontsize=7)

def plot_ax_2(ax, y1, y2, color1, color2, group1, group2, species):
    ax.scatter([0] * len(y1), y1, color=color1, label=group1, alpha=0.8)
    ax.scatter([1] * len(y2), y2, color=color2, label=group2, alpha=0.8)

    
    # Add the means as horizontal lines
    ax.plot([0 - 0.2, 0 + 0.2], [np.log10(np.nanmedian(10 ** y1 - 1) + 1), np.log10(np.nanmedian(10 ** y1 - 1) + 1)], color=color1, lw=2)
    ax.plot([1 - 0.2, 1 + 0.2], [np.log10(np.nanmedian(10 ** y2 - 1) + 1), np.log10(np.nanmedian(10 ** y2 - 1) + 1)], color=color2, lw=2)

    # Add horizontal gridlines at 0.5 intervals
    for y in np.arange(0, max(6, max(max(y1) + 0.5, max(y2) + 0.5)), 1):  # Adjust range to match y-axis limits
        ax.axhline(y, color='lightgray', linestyle='--', linewidth=0.5, zorder=0)

    # Customize the x-axis
    ax.set_xticks([0, 1])
    ax.set_xticklabels([group1, group2])
    ax.set_xlim(-0.5, 1.5)

    # Set axis limits and labels
    ax.set_ylim(0, max(6, max(max(y1) + 0.5, max(y2) + 0.5)))
    ax.set_ylabel('log$_{10}$ counts')

    # Remove the x and y axis lines
    ax.spines['top'].set_visible(False)
    ax.spines['right'].set_visible(False)

    # Remove ticks on the y-axis
    ax.yaxis.set_ticks_position('none')

    ax.set_title(species, fontsize=10, pad=20)

    # Set lighter grid aesthetics
    ax.grid(False)


def plot_ax_4(ax, yHC, ySP, yRR, yCTRL, species):
    ax.scatter([0] * len(yHC), yHC, color='#648FFF', label='HC', alpha=0.8)
    ax.scatter([1] * len(ySP), ySP, color='#DC267F', label='SP', alpha=0.8)
    ax.scatter([2] * len(yRR), yRR, color='#FE6100', label='RR', alpha=0.8)
    ax.scatter([3] * len(yCTRL), yCTRL, color='#848484', label='CTRL', alpha=0.8)

    # Add the means as horizontal lines
    ax.plot([0 - 0.2, 0 + 0.2], [np.nanmean(yHC), np.nanmean(yHC)], color='#648FFF', lw=2)
    ax.plot([1 - 0.2, 1 + 0.2], [np.nanmean(ySP), np.nanmean(ySP)], color='#DC267F', lw=2)
    ax.plot([2 - 0.2, 2 + 0.2], [np.nanmean(yRR), np.nanmean(yRR)], color='#FE6100', lw=2)
    ax.plot([3 - 0.2, 3 + 0.2], [np.nanmean(yCTRL), np.nanmean(yCTRL)], color='#848484', lw=2)

    # Add horizontal gridlines at 0.5 intervals
    for y in np.arange(0, max(5, max(max(yHC) + 0.5, max(ySP) + 0.5, max(yRR) + 0.5, max(yCTRL) + 0.5)), 1):  # Adjust range to match y-axis limits
        ax.axhline(y, color='lightgray', linestyle='--', linewidth=0.5, zorder=0)

    # Customize the x-axis
    ax.set_xticks([0, 1, 2, 3])
    ax.set_xticklabels(['HC', 'SP', 'RR', 'CTRL'])
    ax.set_xlim(-0.5, 3.5)

    # Set axis limits and labels
    ax.set_ylim(0, max(5, max(max(yHC) + 0.5, max(ySP) + 0.5, max(yRR) + 0.5, max(yCTRL) + 0.5)))
    ax.set_ylabel('log$_{10}$ counts')

    # Remove the x and y axis lines
    ax.spines['top'].set_visible(False)
    ax.spines['right'].set_visible(False)

    # Remove ticks on the y-axis
    ax.yaxis.set_ticks_position('none')

    ax.set_title(species, fontsize=10)

    # Set lighter grid aesthetics
    ax.grid(False)

df_normpipe_discarded_common

def plot_species(df, ncols_max=6):
    nplots = len(df)
    nrows = nplots//ncols_max 
    nrows += (nplots % ncols_max != 0)

    if nplots:
        # Create the figure and axis
        fig, axs = plt.subplots(nrows, ncols_max, figsize=(2 * ncols_max, 2 * nrows))

        for i, species in enumerate(df.index):
            ax = axs.ravel()[i] if nplots > 1 else axs
            yHC = np.log10(df.loc[species][['HC1', 'HC2', 'HC3', 'HC4']].astype(float).values + 1)
            ySP = np.log10(df.loc[species][['SP1', 'SP2', 'SP3', 'SP4']].astype(float).values + 1)
            yRR = np.log10(df.loc[species][['RR1', 'RR2', 'RR3', 'RR4']].astype(float).values + 1)
            yCTRL = np.log10(df.loc[species][['ACIDOLA', 'BLACTIS']].astype(float).values + 1)
            
            plot_ax_4(ax, yHC, ySP, yRR, yCTRL, species)
            df_annot = pd.DataFrame({'group': ['HC'] * 4 + ['SP'] * 4 + ['RR'] * 4 + ['CTRL'] * 2, 
                                     'exp': df.loc[species, ['HC1', 'HC2', 'HC3', 'HC4', 'SP1', 'SP2', 'SP3', 'SP4', 'RR1', 'RR2', 'RR3', 'RR4', 'ACIDOLA', 'BLACTIS']].astype(float).values})
            # add_medians(df_annot, ax, x_col='group', y_col='exp', xsync=0.15, ysync=0)

        if nplots % ncols_max:
            for i in range(nplots, nrows * ncols_max ):
                axs.ravel()[i].axis('off')
        plt.tight_layout()



plot_species(df_normpipe_discarded_common.copy().set_index('name'), ncols_max=6)
for format in ['png', 'tiff']: 
    plt.savefig(f'{RESULTS_DIR}/figures/paper/l2fc_NORM+x_discarded.{format}', dpi=DPI)

plot_species(df_normpipe_discarded_normplus.copy().set_index('name'), ncols_max=7)
for format in ['png', 'tiff']: 
    plt.savefig(f'{RESULTS_DIR}/figures/paper/l2fc_NORMx_discarded.{format}', dpi=DPI)

plot_species(df_normpipe_retained_species.copy().set_index('name'), ncols_max=8)



## Plotting differential species

def plot_diff_sopecies(df_pval, cols_A, cols_B, color1, color2, nrows=2, other_labels = None):
    ncols = len(df_pval) // nrows

    if other_labels is None:
        g1, g2 = cols_A[0][:-1], cols_B[0][:-1]

    if ncols:
        # Create the figure and axis
        fig, axs = plt.subplots(nrows, ncols, figsize=(2 * ncols, nrows * 2))

        for i in range(len(df_pval)):
            df_i = df_pval.iloc[i]
            ax = axs.ravel()[i] if ncols > 1 else axs
            y1 = np.log10(df_i[cols_A].astype(float).values + 1)
            y2 = np.log10(df_i[cols_B].astype(float).values + 1)
            
            species = df_i['name']

            df_annot = pd.DataFrame({'group': [i[:-1] for i in cols_A + cols_B], 'exp': df_i[cols_A + cols_B].astype(float).values})

            plot_ax_2(ax, y1=y1, y2=y2, color1=color1, color2=color2, group1=g1, group2=g2, species=species)
            annotation_pval(df_annot, ax, [g1, g2], x_col='group', y_col='exp')
            add_medians(df_annot, ax, x_col='group', y_col='exp', xsync=0.15, ysync=0)

        plt.tight_layout()

# Plot HC vs RR
df_pval_HCvsRR = pd.read_csv(f'{RESULTS_DIR}/differential_abundance/mode{mode}_S{S}_HCvsRR.tsv', sep='\t')
df_pval_HCvsRR_up = df_pval_HCvsRR[(df_pval_HCvsRR['pval_MW'] < 0.15) & (df_pval_HCvsRR['log2FC'] > 0)].sort_values(by='log2FC').head()
df_pval_HCvsRR_down = df_pval_HCvsRR[(df_pval_HCvsRR['pval_MW'] < 0.15) & (df_pval_HCvsRR['log2FC'] <= 0)].sort_values(by='log2FC').tail()

df_pval_HCvsRR_updown = pd.concat([df_pval_HCvsRR_up, df_pval_HCvsRR_down])
df_pval_HCvsRR_updown = df_pval_HCvsRR_updown.iloc[:len(df_pval_HCvsRR_updown) - len(df_pval_HCvsRR_updown) % 2]


ncols = len(df_pval_HCvsRR_updown) 

plot_diff_sopecies(df_pval_HCvsRR_updown,  ['HC1', 'HC2', 'HC3', 'HC4'], ['RR1', 'RR2', 'RR3', 'RR4'], "#648FFF", "#FE6100", nrows=1)

for format in ['png', 'tiff']: 
    plt.savefig(f'{RESULTS_DIR}/figures/paper/l2fc_HCvRR.{format}', dpi=DPI)




df_pval_HCvsRR_up = df_pval_HCvsRR[df_pval_HCvsRR['pval_MW'] < 0.05].sort_values(by='log2FC').head()
df_pval_HCvsRR_down = df_pval_HCvsRR[df_pval_HCvsRR['pval_MW'] < 0.05].sort_values(by='log2FC').tail()

df_pval_HCvsRR_updown = pd.concat([df_pval_HCvsRR_up, df_pval_HCvsRR_down])

ncols = len(df_pval_HCvsRR_updown) 

plot_diff_sopecies(df_pval_HCvsRR_updown,  ['HC1', 'HC2', 'HC3', 'HC4'], ['RR1', 'RR2', 'RR3', 'RR4'], "#648FFF", "#FE6100", nrows=1)

for format in ['png', 'tiff']: 
    plt.savefig(f'{RESULTS_DIR}/figures/paper/pval_HCvRR.{format}', dpi=DPI)

# Plot HC vs SP

df_pval_HCvsSP = pd.read_csv(f'{RESULTS_DIR}/differential_abundance/mode{mode}_S{S}_HCvsSP.tsv', sep='\t')

df_pval_HCvsSP_up = df_pval_HCvsSP[(df_pval_HCvsSP['pval_MW'] < 0.15) & (df_pval_HCvsSP['log2FC'] > 0)].sort_values(by='log2FC').head()
df_pval_HCvsSP_down = df_pval_HCvsSP[(df_pval_HCvsSP['pval_MW'] < 0.15) & (df_pval_HCvsSP['log2FC'] <= 0)].sort_values(by='log2FC').tail()

df_pval_HCvsSP_updown = pd.concat([df_pval_HCvsSP_up, df_pval_HCvsSP_down])
df_pval_HCvsSP_updown = df_pval_HCvsSP_updown.iloc[:len(df_pval_HCvsSP_updown) - len(df_pval_HCvsSP_updown) % 2]

ncols = len(df_pval_HCvsSP_updown)

plot_diff_sopecies(df_pval_HCvsSP_updown,  ['HC1', 'HC2', 'HC3', 'HC4'], ['SP1', 'SP2', 'SP3', 'SP4'], "#648FFF", "#DC267F", nrows=1)

for format in ['png', 'tiff']: 
    plt.savefig(f'{RESULTS_DIR}/figures/paper/l2fc_HCvSP.{format}', dpi=DPI)




df_pval_HCvsSP_up = df_pval_HCvsSP[df_pval_HCvsSP['pval_MW'] < 0.05].sort_values(by='log2FC').head()

ncols = len(df_pval_HCvsSP_up) 

plot_diff_sopecies(df_pval_HCvsSP_up,  ['HC1', 'HC2', 'HC3', 'HC4'], ['SP1', 'SP2', 'SP3', 'SP4'], "#648FFF", "#DC267F", nrows=1)                                        

for format in ['png', 'tiff']: 
    plt.savefig(f'{RESULTS_DIR}/figures/paper/pval_HCvsSP.{format}', dpi=DPI)

# # Plot RR vs SP
# df_pval_RRvsSP = pd.read_csv(f'{RESULTS_DIR}/differential_abundance/mode{mode}_S{S}_RRvsSP.tsv', sep='\t')
# df_pval_RRvsSP_up = df_pval_RRvsSP[(df_pval_RRvsSP['pval_MW'] < 0.15) & (df_pval_RRvsSP['log2FC'] > 0)].sort_values(bsy='log2FC').head(14)
# df_pval_RRvsSP_down = df_pval_RRvsSP[(df_pval_RRvsSP['pval_MW'] < 0.15) & (df_pval_RRvsSP['log2FC'] <= 0)].sort_values(by='log2FC').head(14)

# df_pval_RRvsSP_updown = pd.concat([df_pval_RRvsSP_up, df_pval_RRvsSP_down])
# df_pval_RRvsSP_updown = df_pval_RRvsSP_updown.iloc[:len(df_pval_RRvsSP_updown) - len(df_pval_RRvsSP_updown) % 2]

# ncols = len(df_pval_RRvsSP_updown) // 3

# plot_diff_sopecies(df_pval_RRvsSP_updown,  ['RR1', 'RR2', 'RR3', 'RR4'], ['SP1', 'SP2', 'SP3', 'SP4'],  "#FE6100", "#DC267F", nrows=3)

# for format in ['png', 'tiff']: 
#     plt.savefig(f'{RESULTS_DIR}/figures/paper/l2fc_RRvsSP.{format}', dpi=DPI)


df_pval_RRvsSP_up = df_pval_RRvsSP[df_pval_RRvsSP['pval_MW'] < 0.05].sort_values(by='log2FC').head(16)

ncols = len(df_pval_RRvsSP_up)

plot_diff_sopecies(df_pval_RRvsSP_up,  ['RR1', 'RR2', 'RR3', 'RR4'], ['SP1', 'SP2', 'SP3', 'SP4'],  "#FE6100", "#DC267F", nrows=1) 

for format in ['png', 'tiff']: 
    plt.savefig(f'{RESULTS_DIR}/figures/paper/pval_RRvsSP.{format}', dpi=DPI)

# Plot sex

df_pval_sex = pd.read_csv(f'{RESULTS_DIR}/differential_abundance/mode{mode}_S{S}_sex.tsv', sep='\t')
df_pval_sex = df_pval_sex.rename(columns={'HC1':'Male1', 'HC2':'Male2', 'RR1':'Male3', 'RR2':'Male4', 'SP1':'Male5', 'SP2':'Male6', 'HC3': 'Female1', 'HC4': 'Female2', 'RR3': 'Female3', 'RR4': 'Female4', 'SP3': 'Female5', 'SP4': 'Female6'})

df_pval_sex_up = df_pval_sex[(df_pval_sex['pval_MW'] < 0.15) & (df_pval_sex['log2FC'] > 0)].sort_values(by='log2FC').head(10)
df_pval_sex_down = df_pval_sex[(df_pval_sex['pval_MW'] < 0.15) & (df_pval_sex['log2FC'] <= 0)].sort_values(by='log2FC').tail(10)

df_pval_sex_updown = pd.concat([df_pval_sex_up, df_pval_sex_down])
df_pval_sex_updown = df_pval_sex_updown.iloc[:len(df_pval_sex_updown) - len(df_pval_sex_updown) % 2]

ncols = len(df_pval_sex_updown) 

plot_diff_sopecies(df_pval_sex_updown, [f'Female{i}' for i in range(1,7)], [f'Male{i}' for i in range(1,7)], "#785EF0", "#FFB000", nrows=1)

for format in ['png', 'tiff']: 
    plt.savefig(f'{RESULTS_DIR}/figures/paper/l2fc_sex.{format}', dpi=DPI)


df_pval_sex_up = df_pval_sex[df_pval_sex['pval_MW'] < 0.05].sort_values(by='log2FC').head()

ncols = len(df_pval_sex_up) 

plot_diff_sopecies(df_pval_sex_up,  [f'Male{i}' for i in range(1,7)], [f'Female{i}' for i in range(1,7)], "#785EF0", "#FFB000", nrows=1) 


for format in ['png', 'tiff']: 
    plt.savefig(f'{RESULTS_DIR}/figures/paper/pval_sex.{format}', dpi=DPI)

species = 'Sutterella'

df_all = pd.concat([df_normpipe_retained_species, df_normpipe_discarded_normplus, df_normpipe_discarded_common]).set_index('name')

yHC = np.log10(df_all.loc[species][['HC1', 'HC2', 'HC3', 'HC4']].astype(float).values + 1)
ySP = np.log10(df_all.loc[species][['SP1', 'SP2', 'SP3', 'SP4']].astype(float).values + 1)
yRR = np.log10(df_all.loc[species][['RR1', 'RR2', 'RR3', 'RR4']].astype(float).values + 1)
yCTRL = np.log10(df_all.loc[species][['ACIDOLA', 'BLACTIS']].astype(float).values + 1)

fig, ax = plt.subplots(1, 1)
plot_ax_4(ax, yHC, ySP, yRR, yCTRL, species)


display(df_normpipe_discarded_common[df_normpipe_discarded_common['name'] == species])
display(df_normpipe_discarded_normplus[df_normpipe_discarded_normplus['name'] == species])
display(df_normpipe_retained_species[df_normpipe_retained_species['name'] == species])


display(df_pval_HCvsRR[df_pval_HCvsRR['name'] == species])
display(df_pval_HCvsSP[df_pval_HCvsSP['name'] == species])
display(df_pval_RRvsSP[df_pval_RRvsSP['name'] == species])
display(df_pval_sex[df_pval_sex['name'] == species])

## Doing a heatmap with all properties

df_heatmap = df_all.copy()

df_heatmap['Discarded'] = -1

for name in df_normpipe_discarded_common['name'].values:
    df_heatmap.loc[name, 'Discarded'] = 1

for name in df_normpipe_discarded_normplus['name'].values:
    df_heatmap.loc[name, 'Discarded'] = 2

for name in df_normpipe_retained_species['name'].values:
    df_heatmap.loc[name, 'Discarded'] = 3

df_heatmap['Discarded'] = df_heatmap['Discarded'].astype(str)


df_heatmap['pval'] = -1


columns_of_interest = ['RR1', 'RR2', 'RR3', 'RR4', 'SP1', 'SP2', 'SP3', 'SP4', 'HC1', 'HC2', 'HC3', 'HC4', 'ACIDOLA', 'BLACTIS', 'Discarded']

df_heatmap['mean_counts'] = df_heatmap[columns_of_interest[:-1]].mean(axis=1)
df_heatmap = df_heatmap.sort_values(by='mean_counts', ascending=False)

# Calcular el log10 de los counts para el dataframe ordenado
df_heatmap[columns_of_interest[:-1]] = df_heatmap[columns_of_interest[:-1]].applymap(lambda x: np.log10(x + 1))
df_heatmap = df_heatmap[columns_of_interest]

df_heatmap

# Prepare the figure
fig, axs = plt.subplots(1, 2, figsize=(12.5, 19))

# Define a color map for the Discarded column
discarded_colors = df_heatmap['Discarded'].map({'-1': 'gray', '1': '#d46014', '2': '#e69f08', '3': '#ddcd3d'})

for i, ax in enumerate(axs):
    subset = df_heatmap.iloc[i * (len(df_heatmap) // 2):(i + 1) * (len(df_heatmap) // 2)]
    sns.heatmap(subset.iloc[:, :-1], cmap='Blues', xticklabels=columns_of_interest[:-1], cbar=False, annot=True, ax=ax)

    # Add vertical lines for section separation
    ax.axvline(x=4, color='white', linewidth=2)  # Between RR4 and SP1
    ax.axvline(x=8, color='white', linewidth=2)  # Between SP4 and HC1
    ax.axvline(x=12, color='white', linewidth=2)  # Between HC4 and ACIDOLA

    # Overlay the 'Discarded' labels
    for idx, row in subset.iterrows():
        ax.text(len(columns_of_interest) - 0.5, subset.index.get_loc(idx) + 0.5, row['Discarded'],
                ha='center', va='center', color='black', fontsize=10, weight='normal',
                bbox=dict(facecolor=discarded_colors[idx], edgecolor='#00000000' ))

for ax in axs:
    ax.set_ylabel('')

plt.tight_layout()

for format in ['png', 'tiff']: 
    plt.savefig(f'{RESULTS_DIR}/figures/paper/all_genera_heatmap.{format}', dpi=DPI)

plt.show()



# How does each profiler perform?

We want to plot the distribution of these 151 genera and see how much is profiler contributes to the expression.

rename_mapping = {
        'POOL1': 'RR1', 'POOL2': 'RR2', 'POOL3': 'RR3', 'POOL4': 'RR4',
        'POOL5': 'SP1', 'POOL6': 'SP2', 'POOL7': 'SP3', 'POOL8': 'SP4',
        'POOL9': 'HC1', 'POOL10': 'HC2', 'POOL11': 'HC3', 'POOL12': 'HC4'
    }

list_genera = df_all.index

for pool, sample in rename_mapping.items():
    counts_file = f'{RESULTS_DIR}/summary/{pool}_pass2_mode{mode}_taxgenus_S{S}_NORM+.diversity.tsv'

    # Load the data
    df_counts = pd.read_csv(counts_file, sep='\t').set_index('name')
    available_genera = [i for i in list_genera if i in df_counts.index]
    df_counts = df_counts.loc[available_genera, [f'{prof}_norm' for prof in LIST_PROFILERS] + ['mean_norm']]

list_genera = df_all.index

# Initialize figure with subplots
fig, axs = plt.subplots(2, len(rename_mapping)//2, figsize=(15, 6), sharey=True)

axs = axs.ravel()

for idx, (pool, sample) in enumerate(rename_mapping.items()):
    counts_file = f'{RESULTS_DIR}/summary/{pool}_pass2_mode{mode}_taxgenus_S{S}_NORM+.diversity.tsv'

    # Load the data
    df_counts = pd.read_csv(counts_file, sep='\t').set_index('name')
    available_genera = [i for i in list_genera if i in df_counts.index]
    df_counts = df_counts.loc[available_genera, [f'{prof}_norm' for prof in LIST_PROFILERS] + ['mean_norm']]

    # Prepare data for boxplot
    df_melted = df_counts.melt(var_name='Profiler', value_name='Abundance')
    df_melted['Profiler'] = [i.replace('_norm', '').replace('kraken2', 'KR2').replace('krakenuniq', 'KRU')[:3].upper() for i in df_melted['Profiler']]
    df_melted.loc[df_melted['Profiler'] == 'MEA', 'Profiler'] = 'MEAN'
    df_melted['log(Abundance)'] = np.log10(df_melted['Abundance'] + 1)

    # Generate boxplot
    p = sns.stripplot(x='Profiler', y='log(Abundance)', data=df_melted, ax=axs[idx], size=1, )


    for i, mean in enumerate(df_counts.median(0)):
        axs[idx].plot([i - 0.2, i + 0.2], [np.log10(1+mean), np.log10(1+mean)], color="#232323", zorder=10)

    axs[idx].set_title(f'{sample}')
    axs[idx].set_xlabel('')
    axs[idx].set_ylabel('log$_{10}$(counts + 1)')
    axs[idx].tick_params(axis='x')

    axs[idx].spines['top'].set_visible(False)
    axs[idx].spines['right'].set_visible(False)
plt.tight_layout()

for format in ['png', 'tiff']: 
    plt.savefig(f'{RESULTS_DIR}/figures/paper/figIA.{format}', dpi=DPI)


plt.show()

list_genera = df_all.index

# Initialize figure with subplots
fig, axs = plt.subplots(2, len(rename_mapping)//2, figsize=(15, 6), sharey=True)

axs = axs.ravel()
for idx, (pool, sample) in enumerate(rename_mapping.items()):
    counts_file = f'{RESULTS_DIR}/summary/{pool}_pass2_mode{mode}_taxgenus_S{S}_NORM+.diversity.tsv'

    # Load the data
    df_counts = pd.read_csv(counts_file, sep='\t').set_index('name')
    available_genera = [i for i in list_genera if i in df_counts.index]
    df_counts = df_counts.loc[available_genera, [f'{prof}_norm' for prof in LIST_PROFILERS] + ['mean_norm']]

    for profiler in LIST_PROFILERS:
        df_counts[f'{profiler}_ratio'] = df_counts[f'{profiler}_norm'] / df_counts['mean_norm']

    df_counts = df_counts.loc[:, [f'{prof}_ratio' for prof in LIST_PROFILERS]]



    # Prepare data for boxplot
    df_melted = df_counts.melt(var_name='Profiler', value_name='Ratio')
    df_melted['Profiler'] = [i.replace('_ratio', '').replace('kraken2', 'KR2').replace('krakenuniq', 'KRU')[:3].upper() for i in df_melted['Profiler']]

    # Generate boxplot
    p = sns.stripplot(x='Profiler', y='Ratio', data=df_melted, ax=axs[idx], size=1, )
    
    for i, mean in enumerate(df_counts.median(0)):
        axs[idx].plot([i - 0.2, i + 0.2], [mean, mean], color="#232323", zorder=10)

    axs[idx].axhline(1, color='#bcbcbc', linewidth=1.5, ls='--' )

    axs[idx].set_title(f'{sample}')
    axs[idx].set_xlabel('')
    axs[idx].set_ylabel('Ratio (against average)')


    axs[idx].tick_params(axis='x')

    axs[idx].spines['top'].set_visible(False)
    axs[idx].spines['right'].set_visible(False)

plt.tight_layout()


for format in ['png', 'tiff']: 
    plt.savefig(f'{RESULTS_DIR}/figures/paper/figIA2.{format}', dpi=DPI)


plt.show()

list_genera = df_all.index

# Initialize figure with subplots
fig, axs = plt.subplots(2, len(rename_mapping)//2, figsize=(20, 7), sharey=True)

axs = axs.ravel()

for idx, (pool, sample) in enumerate(rename_mapping.items()):
    counts_file = f'{RESULTS_DIR}/summary/{pool}_pass2_mode{mode}_taxgenus_S{S}_NORM+.diversity.tsv'

    # Load the data
    df_counts = pd.read_csv(counts_file, sep='\t').set_index('name')
    available_genera = [i for i in list_genera if i in df_counts.index]
    df_counts = df_counts.loc[available_genera, [f'{prof}_norm' for prof in LIST_PROFILERS] + ['mean_norm']]
    df_counts.columns = ['CEN', 'GAN', 'KAI', 'KR2', 'KRU', 'MEAN']
    corr_matrix = df_counts.corr(method='spearman')

    # Generate heatmap
    sns.heatmap(corr_matrix, annot=True, cmap='Blues', ax=axs[idx], fmt=".2")
    axs[idx].set_title(f'{sample}')

    axs[idx].set_xticklabels(axs[idx].get_xticklabels(), rotation=0)

plt.tight_layout()

for format in ['png', 'tiff']: 
    plt.savefig(f'{RESULTS_DIR}/figures/paper/figIB.{format}', dpi=DPI)


plt.show()


list_genera = df_all.index

df_counts_all = pd.DataFrame(columns=['CEN', 'GAN', 'KAI', 'KR2', 'KRU', 'MEAN', 'SPECIES'])
for idx, (pool, sample) in enumerate(rename_mapping.items()):
    counts_file = f'{RESULTS_DIR}/summary/{pool}_pass2_mode{mode}_taxgenus_S{S}_NORM+.diversity.tsv'

    # Load the data
    df_counts = pd.read_csv(counts_file, sep='\t').set_index('name')
    available_genera = [i for i in list_genera if i in df_counts.index]
    df_counts = np.log10(df_counts.loc[available_genera, [f'{prof}_norm' for prof in LIST_PROFILERS] + ['mean_norm']] + 1)
    df_counts = df_counts.reset_index()
    
    df_counts.columns = ['SPECIES', 'CEN', 'GAN', 'KAI', 'KR2', 'KRU', 'MEAN']
    df_counts_all = pd.concat([df_counts_all, df_counts])

df_counts_all = df_counts_all.reset_index(drop=True)

# Crear el PairGrid
g = sns.PairGrid(df_counts_all.iloc[:-1, :])

# Aplicar el scatterplot y la línea de regresión fuera de la diagonal
g.map_offdiag(sns.regplot, scatter_kws={'s': 2, 'color': '#bcbcbc', 'alpha': 0.8}, line_kws={'color': '#bc0000', 'linewidth': 1})

for i in range(6):
    for j in range(6):
        if i != j:
            g.axes[i, j].plot([0.5, 6], [0.5, 6], ls='--')

# Establecer límites de 0 a 6 para x e y en todos los gráficos
for ax in g.axes.flatten():
    ax.set_xlim(0, 6)
    ax.set_ylim(0, 6)
    ax.set_xticks([0,6])
    ax.set_yticks([0,6])

# Eliminar los ejes de la diagonal
for i in range(6):
    g.axes[i, i].set_axis_off()

g.axes[0, 0].set_ylabel('CEN')
g.axes[0, 0].set_ylabel('CEN')

# Ajustar el tamaño de la figura
g.figure.set_size_inches(8, 8)

plt.tight_layout()

for format in ['png', 'tiff']: 
    plt.savefig(f'{RESULTS_DIR}/figures/paper/figIC.{format}', dpi=DPI)

plt.show()
